+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- New dependency: `svgdom-utils`. Some algorithms are moved to this crate.
 
 ## [0.8.1] - 2017-02-01
 ### Added